refactor: Clean up output package and reduce size by 21%
- Remove legacy LipbamlRenderer (182 lines) - not used in production
- Remove global singleton pattern - each Renderer creates its own Engine
- Move templateFuncs to engine.go where it's used
- Update tests to use Renderer instead of LipbamlRenderer
- Add test output comparison scripts for future refactoring safety

Output package reduced from 869 to 688 lines (21% reduction).
All tests pass and output remains functionally identical.
