--- pkg/too/output/renderer_test.go
@@ -1,4 +1,0 @@
-package output
-
-// Tests for renderer functionality have been moved to integration tests
-// since rendering is now primarily template-based
